+++ pkg/clientwrappers/parametercontext/parametercontext.go
@@ -155,7 +155,7 @@ func parameterContextIsSync(
 			if expected.Parameter.Name == param.Parameter.Name {
 				notFound = false
 
-				if (!param.Parameter.Sensitive && expected.Parameter.Value != param.Parameter.Value) ||
+				if (!param.Parameter.Sensitive && *expected.Parameter.Value != *param.Parameter.Value) ||
 					expected.Parameter.Description != param.Parameter.Description {
 
 					return false
@@ -202,7 +202,7 @@ func updateRequestPrepare(
 			if expected.Parameter.Name == param.Parameter.Name {
 				notFound = false
 
-				if (!param.Parameter.Sensitive && expected.Parameter.Value != param.Parameter.Value) ||
+				if (!param.Parameter.Sensitive && *expected.Parameter.Value != *param.Parameter.Value) ||
 					expected.Parameter.Description != param.Parameter.Description {
 					notFound = false
 					parameters = append(parameters, expected)
@@ -247,24 +247,26 @@ func updateParameterContextEntity(parameterContext *v1alpha1.NifiParameterContex
 
 	for _, secret := range parameterSecrets {
 		for k, v := range secret.Data {
+			value := string(v)
 			parameters = append(parameters, nigoapi.ParameterEntity{
 				Parameter: &nigoapi.ParameterDto{
 					Name:        k,
 					Description: "",
 					Sensitive:   true,
-					Value:       string(v),
+					Value:       &value,
 				},
 			})
 		}
 	}
 
 	for _, parameter := range parameterContext.Spec.Parameters {
+		value := parameter.Value
 		parameters = append(parameters, nigoapi.ParameterEntity{
 			Parameter: &nigoapi.ParameterDto{
 				Name:        parameter.Name,
 				Description: parameter.Description,
 				Sensitive:   false,
-				Value:       parameter.Value,
+				Value:       &value,
 			},
 		})
 	}