+++ pkg/pki/certmanagerpki/certmanager_user.go
@@ -48,6 +48,12 @@ func (c *certManager) ReconcileUserCertificate(ctx context.Context, user *v1alph
 
 	if err != nil && apierrors.IsNotFound(err) {
 		// the certificate does not exist, let's make one
+		// check if jks is required and create password for it
+		if user.Spec.IncludeJKS {
+			if err := c.injectJKSPassword(ctx, user); err != nil {
+				return nil, err
+			}
+		}
 		cert := c.clusterCertificateForUser(user, scheme)
 		if err = c.client.Create(ctx, cert); err != nil {
 			return nil, errorfactory.New(errorfactory.APIFailure{}, err, "could not create user certificate")
@@ -64,22 +70,9 @@ func (c *certManager) ReconcileUserCertificate(ctx context.Context, user *v1alph
 		return nil, err
 	}
 
-	if user.Spec.IncludeJKS {
-		if secret, err = c.injectJKS(ctx, user, secret); err != nil {
-			return nil, err
-		}
-	}
-
 	// Ensure controller reference on user secret
 	if err = c.ensureControllerReference(ctx, user, secret, scheme); err != nil {
 		return nil, err
-	}
-
-	// Ensure that the secret is populated with the required values.
-	for k, v := range secret.Data {
-		if len(v) == 0 && k != v1alpha1.CoreCACertKey {
-			return nil, errorfactory.New(errorfactory.APIFailure{}, errors.New("not all secret value populated"), "secret is not ready")
-		}
 	}
 
 	return &pkicommon.UserCertificate{
@@ -89,17 +82,25 @@ func (c *certManager) ReconcileUserCertificate(ctx context.Context, user *v1alph
 	}, nil
 }
 
-// injectJKS ensures that a secret contains JKS format when requested
-func (c *certManager) injectJKS(ctx context.Context, user *v1alpha1.NifiUser, secret *corev1.Secret) (*corev1.Secret, error) {
+// injectJKSPassword ensures that a secret contains JKS password when requested
+func (c *certManager) injectJKSPassword(ctx context.Context, user *v1alpha1.NifiUser) error {
 	var err error
-	if secret, err = certutil.EnsureSecretJKS(secret); err != nil {
-		return nil, errorfactory.New(errorfactory.InternalError{}, err, "could not inject secret with jks")
+	secret := &corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      user.Spec.SecretName,
+			Namespace: user.Namespace,
+		},
+		Data: map[string][]byte{},
 	}
-	if err = c.client.Update(ctx, secret); err != nil {
-		return nil, errorfactory.New(errorfactory.APIFailure{}, err, "could not update secret with jks")
+	secret, err = certutil.EnsureSecretPassJKS(secret)
+	if err != nil {
+		return errorfactory.New(errorfactory.InternalError{}, err, "could not inject secret with jks password")
 	}
-	// Fetch the updated secret
-	return c.getUserSecret(ctx, user)
+	if err = c.client.Create(ctx, secret); err != nil {
+		return errorfactory.New(errorfactory.APIFailure{}, err, "could not create secret with jks password")
+	}
+
+	return nil
 }
 
 // ensureControllerReference ensures that a NifiUser owns a given Secret
@@ -128,12 +129,28 @@ func (c *certManager) getUserSecret(ctx context.Context, user *v1alpha1.NifiUser
 	err = c.client.Get(ctx, types.NamespacedName{Name: user.Spec.SecretName, Namespace: user.Namespace}, secret)
 	if err != nil {
 		if apierrors.IsNotFound(err) {
-			err = errorfactory.New(errorfactory.ResourceNotReady{}, err, "user secret not ready")
-		} else {
-			err = errorfactory.New(errorfactory.APIFailure{}, err, "failed to get user secret")
+			return secret, errorfactory.New(errorfactory.ResourceNotReady{}, err, "user secret not ready")
+		}
+		return secret, errorfactory.New(errorfactory.APIFailure{}, err, "failed to get user secret")
+	}
+	if user.Spec.IncludeJKS {
+		if len(secret.Data) != 6 {
+			return secret, errorfactory.New(errorfactory.ResourceNotReady{}, err, "user secret not populated yet")
+		}
+	} else {
+		if len(secret.Data) != 3 {
+			return secret, errorfactory.New(errorfactory.ResourceNotReady{}, err, "user secret not populated yet")
 		}
 	}
-	return
+
+	for _, v := range secret.Data {
+		if len(v) == 0 {
+			return secret, errorfactory.New(errorfactory.ResourceNotReady{},
+				errors.New("not all secret value populated"), "secret is not ready")
+		}
+	}
+
+	return secret, nil
 }
 
 // clusterCertificateForUser generates a Certificate object for a NifiUser
@@ -141,19 +158,33 @@ func (c *certManager) clusterCertificateForUser(user *v1alpha1.NifiUser, scheme 
 	caName, caKind := c.getCA()
 	cert := &certv1.Certificate{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      user.Name,
-			Namespace: user.Namespace,
+			Name:      user.GetName(),
+			Namespace: user.GetNamespace(),
 		},
 		Spec: certv1.CertificateSpec{
 			SecretName:  user.Spec.SecretName,
 			KeyEncoding: certv1.PKCS8,
-			CommonName:  user.Name,
+			CommonName:  user.GetName(),
+			URISANs:     []string{fmt.Sprintf(pkicommon.SpiffeIdTemplate, c.cluster.Name, user.GetNamespace(), user.GetName())},
 			Usages:      []certv1.KeyUsage{certv1.UsageClientAuth, certv1.UsageServerAuth},
 			IssuerRef: certmeta.ObjectReference{
 				Name: caName,
 				Kind: caKind,
 			},
 		},
+	}
+	if user.Spec.IncludeJKS {
+		cert.Spec.Keystores = &certv1.CertificateKeystores{
+			JKS: &certv1.JKSKeystore{
+				Create: true,
+				PasswordSecretRef: certmeta.SecretKeySelector{
+					LocalObjectReference: certmeta.LocalObjectReference{
+						Name: user.Spec.SecretName,
+					},
+					Key: v1alpha1.PasswordKey,
+				},
+			},
+		}
 	}
 	if user.Spec.DNSNames != nil && len(user.Spec.DNSNames) > 0 {
 		cert.Spec.DNSNames = user.Spec.DNSNames